Refactor to using functional components with hooks
