+++ src/App.js
@@ -1,16 +1,13 @@
-/* eslint-disable react/prefer-stateless-function */
 import './App.css';
 import React from 'react';
 import Calculator from './components/Calculator';
 
-class App extends React.Component {
-  render() {
-    return (
-      <main>
-        <Calculator />
-      </main>
-    );
-  }
+function App() {
+  return (
+    <main>
+      <Calculator />
+    </main>
+  );
 }
 
 export default App;
