+++ src/logic/calculate.js
@@ -1,4 +1,4 @@
-import operate from './Operate';
+import operate from './operate';
 
 function isNumber(item) {
   return !!item.match(/[0-9]+/);
